+++ src/main/java/bob/TaskList.java
@@ -1,4 +1,6 @@
-import Tasks.Task;
+package bob;
+
+import bob.task.Task;
 import java.util.ArrayList;
 
 public class TaskList {
