+++ src/main/java/bob/exception/BobException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package bob.exception;
 
 public abstract class BobException extends Exception {
     public BobException() {
