+++ src/main/java/bob/exception/DirectoryNotFoundException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package bob.exception;
 
 public class DirectoryNotFoundException extends BobException {
     public DirectoryNotFoundException() {
