+++ src/main/java/bob/exception/FileNotFoundException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package bob.exception;
 
 public class FileNotFoundException extends BobException {
     public FileNotFoundException() {
