+++ src/main/java/bob/exception/NoTaskException.java
@@ -1,6 +1,4 @@
-package Exceptions;
-
-import Exceptions.BobException;
+package bob.exception;
 
 public class NoTaskException extends BobException {
     public NoTaskException() {
