+++ src/main/java/bob/task/Event.java
@@ -1,4 +1,4 @@
-package Tasks;
+package bob.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
